+++ tests/create-redirect.test.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const { createRedirect } = require('../src/helpers/create-redirect');
+
+describe('createRedirect', () => {
+  it('works', () => {
+    const fn = createRedirect('/maps/overview/');
+    expect(fn).toEqual(expect.any(Function));
+  });
+});